requirements.txt
tekore==4.0.0
simple-term-menu==1.4.1
pycld3